+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - When comparing a staged file to the git head, the modified version now appears on the right rather than the left
 - In most cases, when opening a temp file, BC will now display a more appropriate name (ex. when comparing to clipboard, the contents of the clipboard will be labeled "Clipboard Contents" instead of the randomly generated file name)
 - Invoking "Compare with last save" by right clicking a tab no longer cycles through all open tabs unless the tab isn't active
+- Temporary files created by the extension are now deleted as soon as they are done being used rather than when VSCode is closed
 
 ## 1.0.5